Move src/device_factory -> src/devices/device_factory
